implememnt also a 1:4 bus-to-pin reverse element

+++ include/value.h
@@ -12,7 +12,7 @@ using json = nlohmann::json;
 struct value {
     int type;
     bool boolval;
-    int intval;
+    uint64_t intval;
     float floatval;
     uint64_t arrayval;
     bool is_error;
